refactor: mv process_msg broker.utils
