+++ docs/docs/SUMMARY.md
@@ -414,14 +414,13 @@ search:
                         - [SubscriberProto](api/faststream/broker/subscriber/proto/SubscriberProto.md)
                     - usecase
                         - [SubscriberUsecase](api/faststream/broker/subscriber/usecase/SubscriberUsecase.md)
-                    - utils
-                        - [process_msg](api/faststream/broker/subscriber/utils/process_msg.md)
                 - types
                     - [PublisherMiddleware](api/faststream/broker/types/PublisherMiddleware.md)
                 - utils
                     - [MultiLock](api/faststream/broker/utils/MultiLock.md)
                     - [default_filter](api/faststream/broker/utils/default_filter.md)
                     - [get_watcher_context](api/faststream/broker/utils/get_watcher_context.md)
+                    - [process_msg](api/faststream/broker/utils/process_msg.md)
                     - [resolve_custom_func](api/faststream/broker/utils/resolve_custom_func.md)
                 - wrapper
                     - call
