+++ faststream/rabbit/subscriber/usecase.py
@@ -14,7 +14,7 @@ from typing_extensions import override
 
 from faststream.broker.publisher.fake import FakePublisher
 from faststream.broker.subscriber.usecase import SubscriberUsecase
-from faststream.broker.subscriber.utils import process_msg
+from faststream.broker.utils import process_msg
 from faststream.exceptions import SetupError
 from faststream.rabbit.helpers.declarer import RabbitDeclarer
 from faststream.rabbit.parser import AioPikaParser
